centralize assumptions in one place
The config structure is deemed to be hardcoded, because we already have
a too high coupling between what we match and the DB format.

This also changes the way that border-explorer is called:
border-explorer border-explorer.db ./wikidata/latest-all.json.bz2

Otherwise, it can be called with no arguments, or just the db name:
border-explorer

In this case it will open "border-explorer.db" (the default file name)
will only do the last phase: geojson files generation.