start using the subclass property to get bigger categories
This way we can get much bigger categories since we have the full tree
of sub-classes.

Problem: we also have many more categories that are "useless"; to work
around that we move the list of banned categories to a separate file,
and greatly expand it, up to the top 1000 items by edges. We also show
which categories were *not* filtered out (as comments), and all
descriptions. It will facilitate updating the list in the future.

This list was generated from an SQL query that took ~ 22min on an M1
laptop; once we banned a few categories, we brought the query time back
under 10s. It could maybe be re-generated with this *untested* fish
loop:

for x in (echo "WITH all_parents(id, nat) AS (
                    SELECT DISTINCT natures.id, natures.nat FROM natures
                    UNION SELECT DISTINCT all_parents.id, sub.parent FROM subclass AS sub
                        JOIN all_parents ON all_parents.nat = sub.id)
                SELECT ap.nat FROM edges as edj, all_parents as ap
                    WHERE edj.a = ap.id AND ap.nat IN (SELECT nat FROM all_parents WHERE id = edj.b)
                    GROUP BY ap.nat ORDER BY COUNT(edj.rowid) DESC LIMIT 1000;" | sqlite3 border-explorer.db) ; curl -sSf "https://www.wikidata.org/w/rest.php/wikibase/v1/entities/items/$x" | jq -r '(.id + "\t" + .labels.en + "\t" + .descriptions.en)' ; end > banned-categories-unfiltered.tsv

In order to achieve sub-class-of, we use recursive SQL queries now,
which are more complex (see above). They also benefit from the list of
banned categories, which we put in an SQL table in addition to keeping
it in memory.

This commit also raises the number of kept items to 600; that's a lot of
categories, hopefully not to hard to navigate.